advanced tasks complete

+++ 0x1E-search_algorithms/100-jump.c
@@ -18,7 +18,7 @@ int jump_search(int *array, size_t size, int value)
 	if (array == NULL)
 		return (-1);
 
-	step = sqrt(size);
+	step = (size_t)sqrt(size);
 	low = 0;
 	high = step;
